Refactor RenamingsStorer
