+++ quarkdown-quarkdoc/src/main/kotlin/com/quarkdown/quarkdoc/dokka/transformers/RenamingsStorer.kt
@@ -1,7 +1,10 @@
 package com.quarkdown.quarkdoc.dokka.transformers
 
-import org.jetbrains.dokka.links.DRI
+import com.quarkdown.quarkdoc.dokka.storage.Renaming
+import com.quarkdown.quarkdoc.dokka.storage.RenamingsStorage
 import org.jetbrains.dokka.model.DFunction
+import org.jetbrains.dokka.model.DParameter
+import org.jetbrains.dokka.model.Documentable
 import org.jetbrains.dokka.plugability.DokkaContext
 
 /**
@@ -12,53 +15,22 @@ import org.jetbrains.dokka.plugability.DokkaContext
 class RenamingsStorer(
     context: DokkaContext,
 ) : QuarkdocDocumentableReplacerTransformer(context) {
+    private fun storeIfRenamed(documentable: Documentable) {
+        val name = getOverriddenName(documentable) ?: return
+        RenamingsStorage[documentable.dri] =
+            Renaming(
+                oldName = requireNotNull(documentable.name),
+                newName = name,
+            )
+    }
+
     override fun transformFunction(function: DFunction): AnyWithChanges<DFunction> {
-        getOverriddenName(function)?.let {
-            PrivateRenamingsStorage.renamings[function.dri] =
-                Renaming(
-                    oldName = function.name,
-                    newName = it,
-                )
-        }
-        function.parameters.forEach { parameter ->
-            getOverriddenName(parameter)?.let {
-                PrivateRenamingsStorage.renamings[parameter.dri] =
-                    Renaming(
-                        oldName = parameter.name!!,
-                        newName = it,
-                    )
-            }
-        }
+        storeIfRenamed(function)
         return super.transformFunction(function)
     }
-}
 
-data class Renaming(
-    val oldName: String,
-    val newName: String,
-)
-
-private object PrivateRenamingsStorage {
-    /**
-     * Renamed function names associated with their address.
-     */
-    val renamings: MutableMap<DRI, Renaming> = mutableMapOf()
-}
-
-/**
- * Storage for the old-new function name pairs.
- * This is a mutable map that is populated by the [RenamingsStorer] transformer.
- */
-object RenamingsStorage {
-    /**
-     * @return the new name for the function with the given DRI, or null if it is not found.
-     */
-    operator fun get(dri: DRI): Renaming? = PrivateRenamingsStorage.renamings[dri]
-
-    /**
-     * Clears the stored renamings.
-     */
-    fun clear() {
-        PrivateRenamingsStorage.renamings.clear()
+    override fun transformParameter(parameter: DParameter): AnyWithChanges<DParameter> {
+        storeIfRenamed(parameter)
+        return super.transformParameter(parameter)
     }
 }
